pyproject.toml
[project]
name = "shop-api"
version = "0.0.1"
description = ""
authors = [
    {name = "sultan",email = "sultankurbanov302@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
